Move parsing and conversion functions to ColorUtils
